Add TPM support for master key management

+++ scripts/start_tpm.sh
@@ -1,6 +1,14 @@
 #!/usr/bin/env bash
-BASE=/home/deenka/repos/system-sec-course/rs-s3k/tpm0
+BASE=$(pwd)/tpm0
 
+if ! [ -x "$(command -v swtpm)" ]; then
+  echo "Error: swtpm is not installed." >&2
+  exit 1
+fi
+
+if [ ! -d "${BASE}" ]; then
+  mkdir -p "${BASE}"
+fi
 
 pkill -9 swtpm 2>/dev/null || true
 mkdir -p "${BASE}"
@@ -11,6 +19,7 @@ swtpm socket \
   --tpmstate dir="${BASE}" \
   --ctrl type=unixio,path="${BASE}/swtpm-sock" \
   --log file="${BASE}/swtpm.log",level=20 \
+  --terminate \
   --daemon 
 
   sleep 0.2
